+++ apps/example/src/components/shared/HintLoading.tsx
@@ -7,7 +7,7 @@ import React from "react"
 import PropTypes from "prop-types"
 import { Stack, Spinner } from "@cloudoperators/juno-ui-components"
 
-const HintLoading = ({ text }) => {
+const HintLoading = ({ text }: any) => {
   return (
     <Stack alignment="center">
       <Spinner variant="primary" />