+++ apps/example/src/mocks/browser.ts
@@ -1,3 +1,6 @@
+/* eslint-disable @typescript-eslint/no-unsafe-call */
+/* eslint-disable @typescript-eslint/no-unsafe-argument */
+/* eslint-disable @typescript-eslint/no-unsafe-member-access */
 /*
  * SPDX-FileCopyrightText: 2024 SAP SE or an SAP affiliate company and Juno contributors
  * SPDX-License-Identifier: Apache-2.0
@@ -8,13 +11,13 @@ import getHandlers from "./getHandlers"
 
 let worker = null
 
-export const startWorker = (options) => {
+export const startWorker = (options: any) => {
   worker = setupWorker(...getHandlers(options))
   return worker.start({
     serviceWorker: {
       url: `${options.endpoint}/mockServiceWorker.js`,
     },
-    onUnhandledRequest(request, print) {
+    onUnhandledRequest(request: any, print: any) {
       const url = new URL(request.url)
       // Do not show unhandled request warning for the following paths
       const listToIgnore = ["/src", "/@", "/node_modules"]