+++ packages/ui-components/src/components/Panel/Panel.test.tsx
@@ -8,62 +8,87 @@ import { render, screen, waitFor } from "@testing-library/react"
 import userEvent from "@testing-library/user-event"
 import { describe, expect, test, vi } from "vitest"
 
-import { Panel } from "./Panel.component"
+import { Panel, PanelProps } from "./Panel.component"
+import { PortalProvider } from "../PortalProvider"
 
 const closedClass = "jn-translate-x-[100%]"
+
+const renderPanel = async (props?: PanelProps, extraProps?: any) => {
+  return waitFor(() =>
+    render(
+      <PortalProvider>
+        <Panel {...(props || {})} {...extraProps} />
+      </PortalProvider>
+    )
+  )
+}
 
 describe("Panel", () => {
   describe("Basic Rendering", () => {
     test("renders a panel", async () => {
-      await waitFor(() => render(<Panel />))
+      await renderPanel()
       expect(screen.getByRole("dialog")).toBeInTheDocument()
       expect(screen.getByRole("dialog")).toHaveClass("juno-panel")
     })
 
     test("renders a closed panel by default", async () => {
-      await waitFor(() => render(<Panel />))
+      await renderPanel()
       expect(screen.getByRole("dialog")).toBeInTheDocument()
       expect(screen.getByRole("dialog")).toHaveClass(closedClass)
     })
 
     test("renders a panel without any props", async () => {
-      await waitFor(() => render(<Panel />))
+      await renderPanel()
       expect(screen.getByRole("dialog")).toBeInTheDocument()
     })
 
     test("renders an opened panel", async () => {
-      await waitFor(() => render(<Panel opened />))
+      await renderPanel({
+        opened: true,
+      })
       expect(screen.getByRole("dialog")).toBeInTheDocument()
       expect(screen.getByRole("dialog")).not.toHaveClass(closedClass)
     })
 
     test("renders a panel with heading", async () => {
-      await waitFor(() => render(<Panel heading="My heading" opened />))
+      await renderPanel({
+        heading: "My heading",
+        opened: true,
+      })
       expect(screen.getByRole("dialog")).toBeInTheDocument()
       expect(screen.getByRole("dialog")).toHaveTextContent("My heading")
     })
 
     test("renders a custom classname", async () => {
-      await waitFor(() => render(<Panel className="my-custom-classname" />))
+      await renderPanel({
+        className: "my-custom-classname",
+      })
       expect(screen.getByRole("dialog")).toBeInTheDocument()
       expect(screen.getByRole("dialog")).toHaveClass("my-custom-classname")
     })
 
     test("renders all props as passed", async () => {
-      await waitFor(() => render(<Panel data-xyz={true} />))
+      await renderPanel(
+        {},
+        {
+          "data-xyz": true,
+        }
+      )
       expect(screen.getByRole("dialog")).toBeInTheDocument()
       expect(screen.getByRole("dialog")).toHaveAttribute("data-xyz")
     })
 
     test("renders a panel with undefined className", async () => {
-      await waitFor(() => render(<Panel className={undefined} />))
+      await renderPanel({
+        className: undefined,
+      })
       expect(screen.getByRole("dialog")).toBeInTheDocument()
     })
   })
 
   describe("Conditional Rendering", () => {
     test("renders a panel with close button by default", async () => {
-      await waitFor(() => render(<Panel />))
+      await renderPanel()
       expect(screen.getByRole("dialog")).toBeInTheDocument()
       expect(screen.getByRole("button")).toBeInTheDocument()
       expect(screen.getByLabelText("close")).toBeInTheDocument()
@@ -72,14 +97,18 @@ describe("Panel", () => {
     })
 
     test("renders a panel without a close button", async () => {
-      await waitFor(() => render(<Panel closeable={false} />))
+      await renderPanel({
+        closeable: false,
+      })
       expect(screen.getByRole("dialog")).toBeInTheDocument()
       expect(screen.queryByRole("button")).not.toBeInTheDocument()
       expect(screen.queryByLabelText("close")).not.toBeInTheDocument()
     })
 
     test("renders a panel without a heading", async () => {
-      await waitFor(() => render(<Panel opened />))
+      await renderPanel({
+        opened: true,
+      })
       expect(screen.getByRole("dialog")).toBeInTheDocument()
       expect(screen.getByRole("dialog")).not.toHaveTextContent("My heading")
     })
@@ -87,21 +116,26 @@ describe("Panel", () => {
 
   describe("Events", () => {
     test("on click on close button closes panel", async () => {
-      await waitFor(() => render(<Panel />))
+      await renderPanel()
       await waitFor(() => userEvent.click(screen.getByRole("button")))
       expect(screen.getByRole("dialog")).toHaveClass(closedClass)
     })
 
     test("on click on close button fires onClose handler as passed", async () => {
       const handleClose = vi.fn()
-      await waitFor(() => render(<Panel onClose={handleClose} />))
+      await renderPanel({
+        onClose: handleClose,
+      })
       await waitFor(() => userEvent.click(screen.getByRole("button")))
       expect(handleClose).toHaveBeenCalledTimes(1)
     })
 
     test("on click on close button when panel is already closed", async () => {
       const handleClose = vi.fn()
-      await waitFor(() => render(<Panel opened={false} onClose={handleClose} />))
+      await renderPanel({
+        opened: false,
+        onClose: handleClose,
+      })
       const button = screen.queryByRole("button")
       await waitFor(() => (button ? userEvent.click(button) : null))
       expect(screen.getByRole("dialog")).toHaveClass(closedClass)
@@ -110,7 +144,9 @@ describe("Panel", () => {
 
     test("double-click on close button fires onClose handler twice", async () => {
       const handleClose = vi.fn()
-      await waitFor(() => render(<Panel onClose={handleClose} />))
+      await renderPanel({
+        onClose: handleClose,
+      })
       await waitFor(() => userEvent.dblClick(screen.getByRole("button")))
       expect(handleClose).toHaveBeenCalledTimes(2)
     })